Rename file "setup_method_body.rs" to "setup_tracked_method_body.rs"
